Improved Weapon Stats, Better Projectile Collision, Added Twirl UI
Reworked weapon stats to correctly be upgrades. Previously, if an upgrade doubled a stat, applying it twice would quadruple it. Now, it'd be +100% each time, only tripling it. Stat upgrades are put in the EquipmentManager as 0.3 for +30%, except for integer stats like pierce (which is just 1 for +1).
Moved EquipmentManager into a prefab.
Twirls are now displayed on the ui (image and possibly layout is temporary).
Projectile collision now only hurts enemies on collision enter (rather than every frame), and projectiles break according to their pierce.
Passive upgrade are correctly limited to level 7, and the levelUpsDone variable is correctly adjusted.